neuralbot/requirements.txt
pyyaml
requests
pwntools
